basic on components chapter done
