+++ src/components/Card.css
@@ -1,0 +1,4 @@
+.card{
+    border-radius: 12px;
+    box-shadow: 0 1px 8px rgba(0, 0, 0, 0.25);
+}
