+++ src/components/Card.js
@@ -1,0 +1,9 @@
+import './Card.css'
+
+function Card(props) {
+    
+    return (
+    <div className={`card ${props.className}`}>{props.children}</div>)
+}
+
+export default Card
