+++ src/components/ExpenseItem.js
@@ -1,25 +1,19 @@
-import './ExpenseItem.css'
+import './ExpenseItem.css';
+import ExpenseDate from './ExpenseDate';
+import Card from './Card';
 
 function ExpenseItem(props) {
 
-    const month = props.date.toLocaleString('en-US', { month: 'long' });
-    const day = props.date.toLocaleString('en-US', { day: '2-digit' });
-    const year = props.date.getFullYear();
-
 
     return (
-    <div className='expense-item'>
-        <div className='expense-item__date'>
-            <div>{month}</div>
-            <div>{year}</div>
-            <div>{day}</div>
-        </div>
+    <Card className='expense-item'>
+        <ExpenseDate date={props.date} />
         <div className='expense-item__description'>
             <h2>{props.title}</h2>
             <div className='expense-item__price'>${props.amount}</div>
         </div>
 
-    </div>);
+    </Card>);
 }
 
 export default ExpenseItem;
