+++ src/components/Expenses.css
@@ -1,0 +1,6 @@
+.expenses{
+    width: 70%;
+    margin: 0 auto;
+    background-color: rgba(0, 0, 0, 0.699);
+    padding: 1rem;
+}
